Revert "Revert "Merge remote-tracking branch 'upstream/master'""
This reverts commit 1f6c4e36ce50bc48ab27c43f53e8ddadf2af024f.

+++ code/game/objects/structures/mineral_doors.dm
@@ -17,7 +17,7 @@
 	max_integrity = 200
 	armor = list("melee" = 10, "bullet" = 0, "laser" = 0, "energy" = 100, "bomb" = 10, "bio" = 100, "rad" = 100, "fire" = 50, "acid" = 50)
 	var/sheetType = /obj/item/stack/sheet/metal
-	var/sheetAmount = 7
+	var/sheetAmount = 10
 	var/openSound = 'sound/effects/stonedoor_openclose.ogg'
 	var/closeSound = 'sound/effects/stonedoor_openclose.ogg'
 	var/obj/item/lock_construct/Lock = null
@@ -165,10 +165,14 @@
 				obj_integrity = CLAMP(obj_integrity + 20, 0, max_integrity)
 		else
 			to_chat(user, "<span class='notice'>[src] doesn't need to be repaired.</span>")
-	else if(istype(I, /obj/item/lock_construct)) /* attempt to add a lock */
+	else if(islock(I)) /* attempt to add a lock */
 		return add_lock(I, user) /* call add_lock proc, so we can disable for airlocks */
-	else if(istype(I, /obj/item/key))
+	else if(iskey(I))
 		return check_key(I, user)
+	else if(istype(I, /obj/item/screwdriver) && state == 1)
+		to_chat(user, "<span class='notice'>You begin to take apart the [name].</span>")
+		if(do_after(user, 60, target = src))
+			deconstruct(TRUE)
 	else if(user.a_intent != INTENT_HARM)
 		return attack_hand(user)
 	else
@@ -176,11 +180,10 @@
 
 
 /obj/structure/mineral_door/crowbar_act(mob/living/user, obj/item/I)
-	if(Lock) /* attempt to pry the lock off */
-		if(Lock.pry_off(user,src))
-			qdel(Lock)
-			Lock = null
-			src.desc = "[initial(desc)]"
+	if(Lock && Lock.pry_off(user,src)) /* attempt to pry the lock off */
+		qdel(Lock)
+		Lock = null
+		src.desc = "[initial(desc)]"
 	return
 
 /obj/structure/mineral_door/proc/check_key(obj/item/key/K, mob/user)
@@ -191,10 +194,9 @@
 		return Lock.check_key(K,user)
 
 /obj/structure/mineral_door/proc/check_locked(mob/user)
-	if(Lock)
-		if(Lock.check_locked())
-			to_chat(user, "[src] is bolted [density ? "shut" : "open"]")
-			return TRUE
+	if(Lock && Lock.check_locked())
+		to_chat(user, "[src] is bolted [density ? "shut" : "open"]")
+		return TRUE
 	return FALSE
 
 /obj/structure/mineral_door/proc/add_lock(obj/item/lock_construct/L, mob/user)
